HAPP-1332: Create FetchVaccinesUseCase
Update FetchMedicationsUseCase.kt
